Begin linking the chat app to the landing page app
- these are two separate apps, since the chat app needs to be its own
things in order to package it properly for desktop, but I also need to
use it in the web app as a demo page. So both apps should integrate
seamlessly on the web.
In order to do this:
- add links to the local and production apps to the .env.local and
.env.production files
- create a local server (separately) for development of the landing page
- add a Home button to this app that will link to the landing page, and
only show it when on the web
- I'll need to fix a display bug in this button where it overlaps with
the settings button - I'll do that next, probably using some sort of
menu bar component